rename from AssertAlmostEqual to AssertWorkspaces
