+++ Framework/PythonInterface/mantid/_testing/AssertWorkspaces.py
