+++ Framework/PythonInterface/mantid/testing.py
@@ -1,3 +1,3 @@
-from mantid._testing.AssertAlmostEqual import assert_almost_equal
+from mantid._testing.AssertWorkspaces import assert_almost_equal
 
 __all__ = ["assert_almost_equal"]
