+++ Framework/PythonInterface/test/python/mantid/_testing/AssertWorkspacesTest.py
