+++ Framework/PythonInterface/test/python/mantid/_testing/CMakeLists.txt
@@ -1,6 +1,6 @@
 # mantid._testing tests
 
-set(TEST_PY_FILES AssertAlmostEqualTest.py)
+set(TEST_PY_FILES AssertWorkspacesTest.py)
 
 check_tests_valid(${CMAKE_CURRENT_SOURCE_DIR} ${TEST_PY_FILES})
